Add files via upload
This text file was saved by the BIGS921 software application. It contains a record of user responses to prompts for use input made by the BIGS921 software application. The record shows the following. The user loaded into program memory data saved to a set of files by the BIGS907 software application. The data consists of a predetermined set of game rules for a blackjack-like game belonging to the Finnish 27 family of table-card games, and data about a set of combination-dependent-playing-strategy-tables.  The text file named "game13i.txt" contains a record of the predetermined set of game rules. The text file named "game13o.txt" contains data about the set of combination-dependent-playing-strategy tables. These text files can be found in the repository named "finnish-27-907docs".