Add files via upload
The BIGS921 software application saved this text file. It contains a total-dependent-basic-strategy table.  Basic strategy is the strategy for the play of the player's hand that loses the least amount to the house in the long run with strictly average luck. The table enables players to use basic strategy while playing a blackjack-like game belonging to the Finnish 27 family of table-card games. The table corresponds to a predetermined set of game rules. A text file named "game13i.txt" contains a record of the predetermined set of game rules. The text file can be found in the "finnish-27-907docs" repository.